Moved all commands into separate files (check this later, may be a bit buggy)
