+++ commands/ping.js
@@ -1,0 +1,7 @@
+module.exports.run = async (client, message, args) => {
+    message.channel.send(`Latency: \`${Math.floor(client.ws.ping)}ms\``);
+};
+    
+module.exports.help = {
+    name: 'ping'
+};
