+++ mekhane.js
@@ -11,14 +11,27 @@
 const Discord = require("discord.js");
 const request = require("request");
 const config = require("./config.json");
+const { promisify } = require('util');
+const readdir = promisify(require('fs').readdir);
+
+var powered = "stale memes"
+
 const embedcolor = 13376265;
 const client = new Discord.Client();
+client.commands = new Map();
+client.config = config
+client.embedcolor = embedcolor
+client.powered = powered
+
 const data = require("./data.json");
+
 const nicknames = ["bonk", "bruh", "NO SPECIAL CHARACTERS", "edgelord", "meme thief", "cringe deluxe", "gen rule 7", "stupid", "standard name", "nickname", "generic preset"]; //add your own if you want
-var powered = "stale memes"
+
 var latest; //stores the latest gamenight message
 
 const help = `${config.prefix}help | `;
+client.help = help
+
 var fmsg = help + "wear a mask";
 
 
@@ -41,6 +54,8 @@ function tryParseJSON(jsonString){ //used for whois, doesn't work without it som
     catch (e) { }
     return false;
 };
+
+client.tryParseJSON = tryParseJSON
 
 function specialname(nick){ //checks if a name is nonstandard
 	if(!nick){
@@ -109,6 +124,8 @@ function verify(discid, chanid) { //verifies discord and roblox through eryn
 	})
 };
 
+client.verify = verify
+
 process.on('unhandledRejection', error => {
 	console.error(`nonfatal error code ${error.code}`)
 	console.error('Unhandled promise rejection:', error); //for debugging, remove if needed
@@ -121,9 +138,22 @@ client.on("guildBanRemove", function(gld, usr) {
 })
 
 client.on("ready", () => {
-  console.log(`Bot has started, server size is ${client.users.size} at the time of launch`); 
-  client.user.setActivity(fmsg);
-});
+	readdir('./commands/', (error, files) => {
+		if (error) throw error;
+		files.forEach(file => {
+		  if (!file.endsWith('.js')) return; // make sure the file is what you are looking for
+		  try {
+				const properties = require(`./commands/${file}`);
+				client.commands.set(properties.help.name, properties);
+		  	} catch (err) {
+				throw err;
+		  }  
+		});
+	});
+
+	console.log(`Bot has started, server size is ${client.users.size} at the time of launch`); 
+	client.user.setActivity(fmsg);
+})
 
 client.on('guildMemberAdd', async member => {
 	if(member.user.bot) return;
@@ -192,291 +222,11 @@ client.on("message", async message => {
   	if(message.author.bot) return; //do not process bot messages
   	if(message.content.indexOf(config.prefix) !== 0) return;
   	const args = message.content.slice(config.prefix.length).trim().split(/ +/g); //command handler
- 	const command = args.shift().toLowerCase();
-	switch(command){
-		case "help":
-			let helpembed = {
-				title: "Mekhane Command list",
-				description: `Key: \`<field>\` is always needed, \`[field]\` is optional\nPrefix: \`${config.prefix}\``,
-				color: embedcolor,
-				footer: {
-				  text: `powered by ${powered}`
-				},
-				fields: [
-				  {
-				  	name: "ping",
-				  	value: "Get latency in milliseconds."
-				  },
-				  {
-				  	name: "random",
-				  	value: "Generate a random number between 1 and 10."
-				  },
-				  {
-				  	name: "hellagay <target>",
-				  	value: "Call someone hella gay." 
-				  },
-				  {
-				  	name: "gamertags <amount>",
-				  	value: "Generate xbox gamertags (Max amount is 10)."
-				  },
-				  {
-				  	name: "verify",
-				  	value: "Verifies if your ROBLOX and Discord accounts are linked, will not work if you didn't verify."
-				  },
-				  {
-					name: "avatar [mention | id]",
-					value: "Gets your avatar, or someone else's."
-				  }
-				]
-			};
-			let faildm = {
-			  title: "Mekhane command list",
-			  description: `Could not send commands into DMs. Perhaps enable them?`,
-			  color: embedcolor,
-			  footer: {
-				  text: `powered by ${powered}`
-			  }
-		  }
-		  client.users.cache.get(message.author.id).createDM().then(dm => {dm.send({embed: helpembed})
-		  .catch(err => { //if unable to dm
-				  if(err.code = "50007"){
-					  message.channel.send({embed: faildm})
-					  client.users.fetch(message.author.id).then(tgt => {
-						  console.error(`Failed to send commands to ${tgt.username}#${tgt.discriminator}`)
-					  })
-				  }
-			  })
-		  })
-		break;
-		case "ping":
-			message.channel.send(`Latency: \`${Math.floor(client.ws.ping)}ms\``);
-		break;
-		case "random":
-			message.channel.send("Number generated: " + Math.floor((Math.random() * 10) + 1));
-		break;
-		case "hellagay": //i forgot i made this and it's too funny to remove
-			let biggay = args.join(" ");
-			if(biggay.includes("<@&") || biggay.includes("@everyone") || biggay.includes("@here")) {
-				message.channel.send("Don't try to mention everyone with the bot. Continue and a warning will be issued.")
-			} else {
-				if(biggay){
-					message.channel.send(`${biggay} is hella gay!`);
-				} else {
-					message.channel.send("Please use something to call hella gay.")
-				}
-			}
-		break;
-		case "say":
-			if(config.whitelist.includes(message.author.id)) {
-				let chan = args[0]
-				message.delete().catch(err=>{})
-				if(isNaN(chan)){
-					message.channel.send(args.join(" "))
-				} else {
-					let msg = args.slice(1).join(" ")
-					let tgt = client.guilds.cache.get(message.guild.id).channels.cache.get(chan)
-					if(tgt){
-						tgt.send(msg)
-					} else {
-						message.channel.send(args.join(" "))
-					}
-				}
-			}
-		break;
-		case "gamertags":
-			function names() {
-				//names
-				 let pref = ["Big", "Cool", "Good", "Pro", "Xx_", "Gamer", "Unturned", "Roblox", "Minecraft", "Lucky", "i"];
-				let name = ["Gaming", "Games", "Boy", "Kid", "Killer"];
-				let suff = ["YT", "_xX", "_", "Hacker"];
-				//number generation
-				let numrand = (Math.floor(Math.random() * 1000));
-				let prefrand = (Math.floor(Math.random() * pref.length));
-				let namerand = (Math.floor(Math.random() * name.length));
-				let suffrand = (Math.floor(Math.random() * suff.length));
-				let full = (pref[prefrand] + name[namerand] + numrand.toString() + suff[suffrand]);
-				return full;
-			}
-			let rep = args[0];
-			if(rep > 10) {
-				message.channel.send("Number too high.");
-			} else if(rep < 1) {
-				message.channel.send("Number too low.")
-			} else if(isNaN(rep) == true) {
-				message.channel.send("Numbers only.")
-			} else {
-				let gtags = "";
-				for(i = 1; i <= rep; i++) {
-					gtags += `${names()}\n`
-				}
-				message.channel.send(`\`${gtags}\``);
-			}
-		break;
-		case "playmsg":
-			let msg = args.join(" ");
-			if(config.whitelist.includes(message.author.id)) {
-				if(msg) {
-					client.user.setActivity(help + msg);
-				message.channel.send("\`PLAY MESSAGE UPDATED\`");
-				message.delete();
-				} else {
-					message.channel.send("\`MESSAGE CANNOT BE EMPTY\`")
-					message.delete()
-				}
-			}
-		break;
-		case "verify":
-			verify(message.author.id, message.channel.id)
-		break;
-		case "whois": //this is literally the most unoptimized part of the whole script
-			let them = args.join(" ");
-		let mnt = message.mentions.members.first();
-		if(mnt){
-			them = mnt.id;
-		} else if (!them){
-			message.channel.send("Please insert a valid user.");
-			return
-		}
-		request({ 
-			method: 'GET', 
-			uri: `https://verify.eryn.io/api/user/${them}`
-		}, 
-		function (error, response, body) {
-			let rbxid = tryParseJSON(body);
-			if(rbxid == false){
-				message.channel.send("Something went wrong. Try again.")
-				return
-			} else if(rbxid.status == "ok") {
-				request({ 
-					method: 'GET', 
-					uri: `https://users.roblox.com/v1/users/${rbxid.robloxId}`
-				}, 
-				function (error, response, body) {
-					let whom = JSON.parse(body);
-					let date = new Date(whom.created);
-					let year = date.getFullYear();
-					let month = date.getMonth()+1;
-					let dt = date.getDate();
-					if (dt < 10) {
-  						dt = '0' + dt;
-					}
-					if (month < 10) {
-  						month = '0' + month;
-					}
-					client.users.fetch(them).then(ftchtgt => {
-						let whoisembed = {
-							title: whom.displayName,
-							url: `https://www.roblox.com/users/${rbxid.robloxId}/profile`,
-							description: whom.description,
-							color: embedcolor,
-							author: {
-								name: `${ftchtgt.username}#${ftchtgt.discriminator}`,
-								icon_url: ftchtgt.avatarURL
-							},
-							footer: {
-								text: `powered by ${powered}`
-							},
-							thumbnail: {
-								url: `https://assetgame.roblox.com/Thumbs/Avatar.ashx?username=${rbxid.robloxUsername}`
-							},
-							fields: [
-								{
-									name: "Date of creation",
-									value: dt +'/'+ month +'/'+ year,
-								}
-							]
-						};
-					message.channel.send({embed: whoisembed})
-					})
-			})} else {
-				let embed404 = {
-					title: "Mekhane Account finder",
-        			description: "Error: Account doesn't seem to be linked.",
-        			color: embedcolor,
-       				footer: {
-            			text: `powered by ${powered}`
-					}
-				}
-				message.channel.send({embed: embed404});
-			}
-		});
-		break;
-		case "avatar":
-			let tgtav;
-		let tgt = args[0];
-		let ment = message.mentions.members.first();
-		if(ment){
-			tgtav = ment.id;
-		} else if (isNaN(tgt) == false){
-			tgtav = tgt
-		} else {
-			tgtav = message.author.id
-		}
-		client.users.fetch(tgtav).then(target => {
-			let avembed = {
-				title: `${target.username}#${target.discriminator}'s avatar`,
-				color: embedcolor,
-				footer: {
-					text: `powered by ${powered}`
-				},
-				image: {
-					url: target.avatarURL({dynamic: true})
-				}
-			};
-			message.channel.send({embed: avembed});
-		}).catch(err => {
-			if(err.code == "10013"){
-				let embed404 = {
-					title: "Mekhane Avatar finder",
-        			description: "Could not find user. Maybe double check?",
-        			color: embedcolor,
-       				footer: {
-            			text: `powered by ${powered}`
-					}
-				}
-				message.channel.send({embed: embed404});
-			}
-		});
-		break;
-		case "pwrmsg":
-			let pwr = args.join(" ");
-			if(config.whitelist.includes(message.author.id)) {
-				if(pwr) {
-					powered = pwr;
-					message.channel.send("\`EMBED FOOTER MESSAGE UPDATED\`");
-					message.delete();
-				} else {
-					message.channel.send("\`MESSAGE CANNOT BE EMPTY\`")
-					message.delete();
-				}
-			}
-		break;
-		case "gamenight":
-			let auth = message.member.roles.cache;
-			if(!(auth.has("AUTH-ID-1") || auth.has("AUTH-ID-2"))) return; // does the caller have authorization? you can add your own role IDs
-			let gnm = args.join(" ")
-			let gncache = client.guilds.cache.get(config.server_id).channels.cache // the channels
-			let gnchat = gncache.get("GN-CHAT-ID") // gamenight chat
-			let gncn = gncache.get("GN-ANN-ID") // gamenight announcement chat
-			if(!gnm){
-				message.channel.send("Please include your gamenight message.")
-			} else if (args.join(" ") == "end") {
-				if(!latest){
-					message.channel.send("No gamenight message found.")
-					return
-				}
-				message.delete()
-				latest.delete()
-				message.channel.send(`Gamenight ended by ${message.author.username}#${message.author.discriminator}`)
-				gnchat.overwritePermissions([{id:config.verified_role, deny:["SEND_MESSAGES"]}])
-			} else {
-				message.delete();
-				gncn.send(`<@&${config.gamenight_role}> ${gnm}`).then(msg => latest = msg)
-				gnchat.overwritePermissions([{id:config.verified_role, allow:["SEND_MESSAGES"]}])
-				message.channel.send(`Gamenight started by ${message.author.username}#${message.author.discriminator}`)
-			}
-		break;
-	}	
+	const command = args.shift().toLowerCase();
+	const cmd = client.commands.get(command);
+
+	if (!cmd) return; // the message is not a command we know of
+	cmd.run(client, message, args); // run the command with client object, message object and args array
 });
 
-client.login(config.token); //log into discord
+client.login(client.config.token); //log into discord
